# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: AgentOps Desktop HTTP API
  description: |
    HTTP REST API for AgentOps Desktop — manage AI agents, goals, tasks, schedules, squads, cost budgets, adapters, and skills.

    **Base URL:** `http://localhost:3967`
    **Authentication:** Bearer token via `POST /auth/login`. Include as `Authorization: Bearer <token>` on all `/api/*` routes.
    **Default port:** 3967 (configurable via `PORT` env var)
  version: "0.2.0"
  contact:
    name: AgentOps Team
  license:
    name: UNLICENSED
servers:
  - url: "http://localhost:3967"
    description: Local HTTP API server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Session token from POST /auth/login
  schemas:
    Error:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: NOT_FOUND
            message:
              type: string
              example: Resource not found
            status:
              type: integer
              example: 404
    Agent:
      type: object
      properties:
        id:
          type: string
          example: "agent-1"
        name:
          type: string
          example: "claude-dev"
        type:
          type: string
          enum: [autonomous, supervised, manual]
          example: autonomous
        status:
          type: string
          enum: [idle, running, paused, error]
          example: idle
        command:
          type: string
          nullable: true
        execPath:
          type: string
          nullable: true
          example: /usr/local/bin/claude
        cwd:
          type: string
          nullable: true
          example: /project
        createdAt:
          type: integer
          example: 1716900000000
        updatedAt:
          type: integer
          example: 1716900000000
...
paths:
...
  /api/agents:
    get:
      summary: List agents
      operationId: agentsList
      tags: [Agents]
      security:
        - bearerAuth: []
      parameters:
        - name: offset
          in: query
          schema: { type: integer, minimum: 0 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100 }
        - name: status
          in: query
          schema: { type: string, enum: [idle, running, paused, error] }
        - name: sortBy
          in: query
          schema: { type: string, enum: [name, status, createdAt, updatedAt] }
        - name: sortOrder
          in: query
          schema: { type: string, enum: [asc, desc] }
      responses:
        "200":
          description: Agent list
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Agent" }
    post:
      summary: Create agent
      operationId: agentsCreate
      tags: [Agents]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 200 }
                type: { type: string, enum: [autonomous, supervised, manual] }
                command: { type: string, maxLength: 1000 }
                execPath: { type: string, maxLength: 1000 }
                cwd: { type: string, maxLength: 500 }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data: { $ref: "#/components/schemas/Agent" }
        "422":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/agents/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      summary: Get agent
      operationId: agentsGet
      tags: [Agents]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Agent
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data: { $ref: "#/components/schemas/Agent" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    patch:
      summary: Update agent
      operationId: agentsUpdate
      tags: [Agents]
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                type: { type: string, enum: [autonomous, supervised, manual] }
                status: { type: string, enum: [idle, running, paused, error] }
                command: { type: string }
                execPath: { type: string }
                cwd: { type: string }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data: { $ref: "#/components/schemas/Agent" }
        "404":
          description: Not found
    delete:
      summary: Delete agent
      operationId: agentsDelete
      tags: [Agents]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Deleted
        "404":
          description: Not found

  # ── Goals ──
...
  /api/cost/budgets/agent/{agentId}:
    parameters:
      - { name: agentId, in: path, required: true, schema: { type: string } }
    get:
      summary: Get budget by agent
      operationId: costBudgetsByAgent
      tags: [Cost]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Budget
        "404":
          description: Not found
...
  /api/adapter-registry/search:
    get:
      summary: Search adapters
      operationId: adapterRegistrySearch
      tags: [Adapter Registry]
      security:
        - bearerAuth: []
      parameters:
        - { name: q, in: query, required: true, schema: { type: string, minLength: 1 } }
        - { name: remote, in: query, schema: { type: boolean } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
      responses:
        "200":
          description: Search results
  /api/adapter-registry/installed:
    get:
      summary: List installed packages
      operationId: adapterRegistryInstalled
      tags: [Adapter Registry]
      security:
        - bearerAuth: []
      parameters:
        - { name: offset, in: query, schema: { type: integer } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { name: source, in: query, schema: { type: string } }
      responses:
        "200":
          description: Installed packages
  /api/adapter-registry/featured:
    get:
      summary: Featured adapters
      operationId: adapterRegistryFeatured
      tags: [Adapter Registry]
      security:
        - bearerAuth: []
      parameters:
        - { name: limit, in: query, schema: { type: integer } }
      responses:
        "200":
          description: Featured adapters
  /api/adapter-registry/updates:
    get:
      summary: Check for updates
      operationId: adapterRegistryUpdates
      tags: [Adapter Registry]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Available updates
  /api/adapter-registry/scan:
    get:
      summary: Scan local adapters
      operationId: adapterRegistryScan
      tags: [Adapter Registry]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Discovered adapters
  /api/adapter-registry/package/{name}:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
    get:
      summary: Get package details
      operationId: adapterRegistryPackage
      tags: [Adapter Registry]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Package details
        "404":
          description: Not found
  /api/adapter-registry/install:
    post:
      summary: Install from remote
      operationId: adapterRegistryInstall
      tags: [Adapter Registry]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                version: { type: string }
                autoLoad: { type: boolean }
      responses:
        "201":
          description: Installed
        "422":
          description: Install error
  /api/adapter-registry/install-file:
    post:
      summary: Install from local file
      operationId: adapterRegistryInstallFile
      tags: [Adapter Registry]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [filePath]
              properties:
                filePath: { type: string }
                name: { type: string }
                autoLoad: { type: boolean }
      responses:
        "201":
          description: Installed
        "422":
          description: Install error
  /api/adapter-registry/register-local:
    post:
      summary: Register local adapter
      operationId: adapterRegistryRegisterLocal
      tags: [Adapter Registry]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [discovered]
              properties:
                discovered: { type: object }
      responses:
        "201":
          description: Registered
        "422":
          description: Registration error
  /api/adapter-registry/uninstall:
    post:
      summary: Uninstall adapter
      operationId: adapterRegistryUninstall
      tags: [Adapter Registry]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                removeFiles: { type: boolean }
      responses:
        "200":
          description: Uninstalled
        "422":
          description: Uninstall error
  /api/adapter-registry/update:
    post:
      summary: Update adapter
      operationId: adapterRegistryUpdate
      tags: [Adapter Registry]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                version: { type: string }
      responses:
        "200":
          description: Updated
        "422":
          description: Update error

  # ── Skills ──
...